Released theme switcher status
