+++ app/src/main/java/com/example/playlistmaker/MainActivity.kt
@@ -9,6 +9,11 @@ class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
+
+        val sharedPrefsTheme = getSharedPreferences(THEME_SWITCHER_VALUE, MODE_PRIVATE)
+        var themeMode = sharedPrefsTheme.getBoolean(THEME_SWITCHER_KEY, false)
+        (applicationContext as App).switchTheme(themeMode)
+
         val findButton = findViewById<Button>(R.id.find_button)
         findButton.setOnClickListener {
             val displayIntent = Intent(this, SearchActivity::class.java)
